94/08/01 ver. 0.53.8 unofficial pre-release
Mon Aug 01 1994 Jmaclean

-	Added new parser combo lexer.l/parser.y from Scott for all
	to play with. A bit rough but Scott would enjoy feedback :-).
- 	Fix for bug in cursor positioning that causes garbage sometimes
  	This can be a MAJOR improvement to some people because it's appliation-
  	specific.
- 	Invisible cursor support in console mode.  More visually pleasing.
- 	Cursor in ANSI and NCURSES is now put at position 0,0 whenever the
  	cursor is invisible or off-screen.
- 	Improvements to character blinking disable/enable code (this stops the
  	annoying blinking when you run PCTools over a telnet line)

-	Note from Markkk about new ansi_xterm :

"ANSI Xterm 2.0 is a specially modified Xterm that is compatible with DOSEMU
and to enable the use of all 16 colors, and the entire PC character set.
This makes DOSEMU just like xdos in this area!

Furthermore, just follow the instructions included with ANSI Xterm 2.0
for an easy compile.  Then all you need to do is type "xdosemu" to bring
up an Xwindow running DOSEMU.   The xdosemu command automatically detects
for the existence of ansi_xterm.  If it does not exist, it detects for
a regular xterm and rxvt next.

Try it, and let Mark Rejhon know at mdrejhon@csclub.uwaterloo.ca
as he is responsible for this."