First (yet mostly unsuccessfull) attempt on uncommitted memory support:  - Virtual Support extension added  - DPMI functions 0x506 and 0x507 added  - All the necessary functions for memory requesting/mapping added  - Functions for changing memory attributes added.
Uncommited memory is being allocated with an anonymous mmap(), and therefore
can not be aliased (not from SHM pool). "Uncommittness" is done via the
PROT_NONE protection, which guarantees a page fault on access.

But only a few programs are working properly with that implementation right
now. It definitely requires more work. I am committing it regardless - otherwise
it will simply never be started:)


git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@325 7758b857-130e-0410-ab9c-d8735a528d18