From: dcoffin@cybercom.net To: bart@dosemu.org Subject: Incorrect bitshifts in instremu.c
Hi Bart,

     My fiancee really loves playing her old DOS games
on DOSEMU, so thanks for the hard work!

     Running my Scrabble game in DOSEMU, I saw visual
defects.  After many hours of debugging, I traced the
problem to instremu.c.

     My code does "shl al,cl", which should set AL to
zero when CL is 8.  But DOSEMU treats the 8 as though
it were a zero.  No x86 CPU behaves this way.

     According to the MASM reference:

"On the 80186-80386 processors, shift counts larger than
31 are masked off, but on the 8088 and 8086, larger shift
counts are performed despite the inefficiency involved."

     I experimentated with the other shift and rotate
commands, and then made a patch to imitate the behavior
of 80186 and later CPUs.  RCL and RCR are especially
interesting...
                                Dave Coffin  6/3/2004


git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@511 7758b857-130e-0410-ab9c-d8735a528d18