Eliminate TrapVgaOn; simplify unmapped VGA memory.
TrapVgaOn is a funny way to detect vgaemu memory mapping; Video->update_screen
does the trick too. Unmapped VGA memory can be mapped RO to the scratch
page with 0xff bytes; writes to it can be trapped (JIT) or avoided (SIM),
but reads to the unmapped memory no longer need to be trapped or simulated
anywhere.