Mouse text cursor screen corruption fix.
This fixes a regression from git commit 699e9564:
When a mouse grab is in effect in a DOSEMU window under X,
and the native text mouse cursor is activated, the text screen
gets corrupted by moving the mouse. This fixes the screen corruption.
Patch modified from Eric Larouche: SF patch 3539932.