Implement EMS map pages and jump.
Based on SF Patch #3537245 (Eric Larouche).
The patch implements EMS 4.0 function 0x55, "alter page map and jump",
which is currently missing in DOSEMU. It has been successfully validated
with a test program under DOS. The test program itself was validated by
running it first with MS-DOS and EMM386. The execution in DOSEMU gives
the same result.