coopth: add some API that aims for a simple cooperation with pthreads.
vga: eliminate signal handling reentry and copy video mem in a
separate thread, as was suggested by Bart in the comments added
by commit ef4638db.

Actually, this is a whole new application of coopthreads_v2: it can
be used as a bridge to pthreads. Without coopthreads, even doing a video
mem copy in a separate thread would not allow to get rid of a signal
handling reentry. Or it will require a gross hacks, including the
int10h calls from the thread, which is probably not practically possible.