fix IF flag during DOS callouts from dosemu inthandlers
When dosemu revect inthandler is executed, it does not disable interrupts.
Non-revect inthandlers did fake_iret() as a first thing, which re-enabled
interrupts.
This was fine when all dosemu inthandlers were plain C.
But they started to do DOS callouts at one point. First, by fake_int_to(),
then with do_call_back (pkt, ipx), and now - with coopthreads. So all the
inthandlers are now doing callouts to vm86, and it is not good to do so
with interrupts enabled.

This patch fixes the problem the following way:
- For non-revect ints it introduces set_iret() which sets CS:IP to hlt
and takes part of flags from stack. When hlt triggers, there is ret_from_int()
which does retf and takes another part of flags.
- For revect ints there is rvc_int_pre()/rvc_int_post() that takes care
of the flags on behalf of coopthreads (sorry for another coopthreads-based
solution)

This patch needs a review!!!