Revert "Changed smalloc to use a double-linked list for a little performance"
This reverts commit ff742dbcfee5ffc764b1822eccb8739e14e21e7f.

Conflicts:
	src/base/misc/smalloc.c

We never traverse the smalloc lists backwards more than by one step,
and even that happens only right after traversing it forward.
Therefore arguably it should be faster to just remember the prev ptr.
Was the aforementioned speedup really benchmarked?

Yes, there is an extra mov with the "remember prev ptr" approach,
but it is the mov between registers, while using the "prev" member
is a memory access.
Furthermore, the CPU can optimize away that extra mov completely
by the use of the dynamic register renaming:
http://en.wikipedia.org/wiki/Register_renaming
When we fetch the "next" pointer, its destination is not needed to
be the same physical register as before; it is only a same
"architectural" register. Same is for the "prev" pointer.
Therefore the CPU can just shift its Reorder Buffer for both
registers at once (rotate it by 1) and do just the memory fetch.

IMHO adding an extra link to the list needs a strong reasoning.
The smalloc_fast approach (now seems removed by 387b9ffd792d) is
a proper fix for the traverse slowdown problem anyway.