fossil: got rid of ugly fossil_dr_hook()
Awake reader thread from the irq handler instead.
This needs testing.

The alternative approach could be to integrate the irq handler
into fossil.com... but really, perhaps instead it might be better to
nuke out fossil.com entirely, or, at lease, make it not a TSR. After
all, coopthreads was implemented for exactly that kind of things: to
embed everything into dosemu code, so if the irq handler can be embedded
too - why not?