vgaemu: make LFB mappings work when lfb_base<mem_base.
Since it is always at the same place, use one non-fixed alias mapping
for the LFB at init time, and only adjust memory protections for video mode
sets. Make mprotect_mapping for the LFB work with respect to lfb_base, instead
of mem_base. This makes the LFB safe if it is allocated below mem_base.