EMM: Fix potential array overrun.
The maximum number of EMM handles is defined as 255 and consequently
the handle_info array is sized to have valid values of 0..254. The code
checks for out of bounds values < 0 and > 254 *OR*
handle_info[handle].active == 0, so there is the potential to overrun the
handle_info array if the handle is invalid, this patch corrects that.