cpuemu: extend protected part to int xx instructions.
Ints are always interpreted, but sometimes (e.g. for FPU emulators,
INT 3x) the int instruction is modified by the int handler to
become an actual FPU instruction). This makes sure that then
the whole block is retranslated so it is as long as possible.

This greatly speeds up programs that use a hybrid FPU emulation/real FPU
approach (SF support #264: CPU benchmarks).