coopth: remove no longer needed coopth_leave()
With the latest changes there seems to be no need to protect "left"
threads against cancellation by coopth_flush(). Previously it was
possible to have several threads with same tid, one "left" one "normal",
and the result of that mess was that coopth_flush() was confused about
what is to cancel. But now all such mixes are rooted out and forbidden.
coopth_leave() is now a macro that expands to coopth_detach().
Can later be re-implemented if some difference with coopth_detach()
is found.