disable xf86vm extension by default [fixes #547]
It seems it only causes problems these days.
Lets see if someone wants it back and will object.

There are the very strange artifacts with XResizeWindow() after the
mode switch, and if dosemu crash, the entire desktop is unusable.

The downside is some performance drop and, more importantly, the degraded
graphics quality. Without a mode switch there is no more interpolation
done by monitor. But this should be fixed with the software filtering in
the future.

xrandr should be investigated as an alternative.
But this patch allows to get something functional quickly.