clarify the right to run DOS software under dosemu...
You thought running DOS software under dosemu is illegal, did you? :)

Anyway, this addresses the concerns expressed in this posting:
https://www.mail-archive.com/dosemu-devel@lists.sourceforge.net/msg00328.html
---
The essence of clause 5 is that if your have a weird lawyer that
interprets the GPL this way you are still fine, because we the authors
explicitly say it is not a problem.

Use of a program is not always allowed.  You have to make a copy of a
program to use it, and furthermore it can be considered a perforamnce.
Both things that depending of your exact copyright law are regulated.
---