change default license to "GPLv2 or later"
The GPL faq says:
---
While you may release your project (either your original work and/or work
that you received and modified) under GPLv2-or-later in this case, note
that the other code you're using must remain under GPLv2 only. As long as
your project depends on that code, you won't be able to upgrade the license
of your project to GPLv3-or-later, and the work as a whole (any combination
of both your project and the other code) can only be conveyed under the term
of GPLv2.
---