coopth: re-implement coopth_leave() separately from coopth_detach() - again
coopth_leave() should be atomic, while coopth_detach() not.
coopth_detach() should have a separate entry point and switching abilitied,
coopth_leave() not.
It is better to have them parted forever.

Since coopth_leave() is now atomic and the resulting thread is
non-switchable, the considerations of commit a027b6d7a8 do not apply too:
the attempt to cancel the left thread is simply not possible to make
as such a thread is always current.

This replaces the fix of 551371689