Add XRandR support; rework input grabbing
XRandR support is fully added and works properly in a multihead configuration
if the user's version is at least 1.2.  It works basically like the existing
XVidMode code and is preferred whenever it is available.  It will also properly
restore the dosemu window's position when exiting fullscreen mode.

Keyboard and mouse grabbing had to be reworked to account for multihead (e.g.,
fullscreen DOS window on one output but with the ability to use applications on
a different output).