Add some compile-time optional log annotations.
Add the ability to annotate each log entry with the VM instruction pointer for
easier debugging of emulated hardware states.  Also add the ability to search
the backtrace to flag whether the log entry appeared to be caused by the DOS
program (to be able to tell it apart from events caused by a dosemu thread or
SIGALRM).