revert part of eedef9a2a that unblocks some signals earlier
It is absolutely unclear why this was needed. I should write
better log messages next time...
OTOH now it looks unsafe to unblock signals early because the
threads will inherit the sigmask and allow SIGIO processing not
in main thread. This may result in not returning from vm86() by
SIGIO, for example.