Install: add error checking on asprintf / malloc
Adds the checks for failed mallocs which should silence the compiler
whinges.

Also noticed there was a corner case where the user supplied an absolute
path, when asked to confirm this, changed it to a relative one. Previously
this would preceed without the value of getenv('PWD') being prepended.
This was probably benign, however it was a different symantic to if the
user specified a relative path initially.