signal: restore es and ds when returning to dosemu code
Also make sure ss is corrent in sigcontext before sigreturn().

This all is not strictly necessary, but makes code easier to
understand, and also its good to have the correct segreg values
when returning back to dosemu's 64bit code: incorrect ds/es values
do not matter but can't be re-loaded, so better to stay safe.