Update asm/ldt.h from kernel 4.1 uapi headers
Update our local copy of asm/ldt.h as upstream now matches our type
definition for base_addr. Ideally we'd cull it and use the system copy,
but upstream only made the change for v2.6.24 so it's a little early to
drop support for prior versions.

Update the references to use the struct's new name.