dpmi: use dynamic reloc in both PIC and non-PIC mode
The "pure-PIC" code is disabled and left just as an example.
Dynamic relocs are allowed in PIC mode, and in fact, without this patch:
https://sourceware.org/ml/binutils/2015-02/msg00091.html
it is not even possible to get a warning on them (with the patch the
warning may be possible, but again, regardless of the PIC mode in use).

So I am absolutely not sure what a big deal was it in dosemu1,
see dpmi_xfr_buffer mess in dosemu1 code.