Add initial support for KVM.
This is sufficient to boot DOS with $_cpu_emu=(0) and $_hogthreshold=(0),
but without redirector support. The main issue is that INTs are not
trapped so we can only have HLT and I/O traps.

I have only tested this so far in basic vm86 text mode, no DPMI or graphics,
but it's promising because the speed is comparable to vm86() syscall speed,
and vm86() seems to be on life support these days.