Automatically redirect drives again.
The strategy is to disable redirecting when DOS tries to open
config.sys; after opening config.sys it will often try to reinitialize
the int2x vectors, and this way this can be caught in int13.

An alternative approach would be to not hook any ints before DOS
tries to open config.sys but for some reason Win98 DOS does not like
this.