disks: Allow the user to force disk type at runtime
This patch adds a config option that allows the user to choose a predetermined
disk type at runtime for hdimage that create a fatfs filesystem from a
directory. This is useful with older versions of DOS e.g. those before v4.0
that did not support large disks or FAT16B filesystems. The disk geometry
chosen is that used by the early IBM BIOS. This page details the types
available:
http://webpages.charter.net/danrollins/techhelp/0054.HTM

Note: Currently it is not possible to tag the default value like:
$_hdimage = "drives/*:hdtype1"
but they may be specified individually as
$_hdimage = "drives/c:hdtype1 drives/d:hdtype2 drives/e"