Add initial support for KVM in V86 mode (with VME enabled).
This implementation is almost 100% compatible with the vm86 syscall,
including INT redirection, I/O traps via GPFs, and flag handling with
IOPL=0. CPU exceptions are trapped using a mini V86 monitor that
pushes the state and executes HLT, which is in turn trapped by KVM.

Even though we have a double trap the advantage vs. real mode in KVM
is that only one syscall is necessary since the state is on the stack.

I still have to check ioperm(), and exceptions other than 1, 3, or 0xd.