dosdebug: Tidy up process selection
Prior to this patch stale debug pipes would need to be deleted before
auto detection of a single dosemu process could succeed. Auto detection
now ignores any pipes for which a process cannot be found. If multiple
dosemu processes are found, the original behaviour of exiting and
listing the matching process IDs is performed.