dosdebug: fix INTxx processing [fixes #136]
We can't process it in pre_vm86() hook because of the very interesting
scenatio: before we get to pre_vm86(), the IRQ may occur and disable
TF. When IRQ returns, TF gets restored, but it does not generate the
trap until the next insn is executed. This is perfectly fine unless
this next insn is INTxx, in which case we can't step into it.
So just move the INTxx processing directly to mhp_trace(): it then
happens before any IRQ.