remove dreaded pragma pack from aspi.h
This was an absolutely horrible bug that made all the structures
incompatible between the files that include aspi.h and the rest of
dosemu.

valgrind was showing only this:

==7844== Conditional jump or move depends on uninitialised value(s)
==7844==    at 0x45FBBD: run_vm86 (do_vm86.c:554)
==7844==    by 0x45FBBD: loopstep_run_vm86 (do_vm86.c:612)
==7844==    by 0x4208AD: main (emu.c:383)

which looks a total rubbish and false-positive.
A very deep and extensive digging required to track that to
this horrible pragma stuff.