simx86: remove VGAEMU call from e_emu_fault for native DPMI code.
Since dpmi_fault() now takes care of VGAEMU faults when cs refers
to a real LDT selector, we don't have to do it in e_emu_fault
any more.

This corrects 1dbc40e5: the only thing we need to check for native
DPMI code is writes to JIT protected pages, all other code is not
relevant, so I moved all native DPMI code to the top of the function
e_emu_fault() with an early return.