dpmi: remove the use of register variable for stack pointer
register unsigned long sp asm("sp")
... is some kind of an extension that is not properly supported by
both gcc and clang.
gcc initializes the variable sp by the value of the stack pointer,
but it doesn't update the value when reading from that variable,
so the stack pointer was offsetted. If declared volatile, then gcc
does the right thing but emits the warning -Wvolatile-register-var.
clang simply warns that the variable is not initialized.
In any case, the gcc manual:
https://gcc.gnu.org/onlinedocs/gcc/Local-Register-Variables.html#Local-Register-Variables
makes such use invalid.

For compatibility with both gcc and clang, just use the inline asm.
There is a better solution in coopth_dpmi1 branch.