Cleanup: Avoid name clash with linux uapi headers
The include file linux/input-event-codes.h defines KEY_* names which we
also define locally. Rename ours from KEY_* to DKY_* to avoid the clash.
This is a necessary cleanup to be able to delete our local copies of
joystick.h and keyboard.h, and so utilise the kernel's uapi headers
instead.