system.com: respect -T even for missing commands [#177]
The old logic is to not terminate if the command is missing.
This allows to see what problem happened.
Unfortunately this doesn't work well with -q, and also "hangs"
the automated dosemu invocations like what openwatcom does.
Now as we correctly pass the exit status, we can report the missing
command with that. Additionally, the error is now printerd into a log.