dpmi: dont switch to client until signals processed [fixes #169]
dpmi_return_request() is called by signals and sets the flag
to return. Unfortunately, by a single return not all signals
are processed, but the flag gets cleared, leaving the stalled
signals in queue. Together with the just fixed instremu bug,
that caused the complete dead-locks or a major slowdowns.

In the past, dosemu used to process all signals at once.
That was actually for the short period of time: since commit
3be6d8f2c and until the signal handling was handed over to
coopth subsystem. It would be untrivial to teach coopth to
process signals at once as it doesn't have the priority system
right now. So the solution is to not run DPMI until all signals
are processed.