vgaemu: fix lfb breakage of commit 43453825 [#204]
After 43453825 the lfb was not mapped because I removed the hack
in register_hardware_ram(), and it appeared that init_hardware_ram()
was called earlier than allocating lfb so it wasn't able to map it
either.
Introduced map_hardware_ram_manual() to do the thing explicitly.
Also use raw lfb pointer for the physical address, rather than the
mem_base'd one. This looks logical and will help kvm_base.