mouse: use separate resolution for mickeys scaling
This fixes a very obscure case when the mouse resolution is much
different from the screen resolution. This mostly happens only when
winos2 mode is active. But since it doesn't use mickeys, the only
way to reproduce this obscure bug was to use
$_mouse_dev = "com1"
$_com1 = "mouse"
which, when grab is enabled, gets routed through direct mickey
movements.

Oh man, this all gets too complicated...