disks: Include sys/types.h for loff_t
Include the sys/types.h header file so that loff_t is defined. Currently
this makes no difference to the build as the users of disks.h seem to
have included sys/types.h for their own purposes. However, if using
disks.h outside of the normal build, for example to use its struct
definitions in a standalone tool, this allows compilation.