disks: MS-DOS 3.10 needs AX set to use our boot code
Set the AX register in the transfer block when booting from our boot
code for OLD MS-DOS and OLD PC-DOS. After disassembling the native boot
block (#278) it can be seen that this value is derived from the offset
of the MSDOS.SYS file within the root directory data. To calculate the
offset we have to assume, as early versions of DOS do, that the
directory is ordered with IO.SYS as the first entry and MSDOS.SYS as
the second. The original boot block does this by taking the size of
IO.SYS and rounding up to the next cluster we can determine the offset
of MSDOS.SYS. We can calculate the offset of the file directly. MS-DOS
v3.10 requires this value, divided by 16, to be passed in the AX
register. Other versions of DOS served by this code (MS-DOS 3.00->3.31
& PC-DOS 3.00->3.31) do not care about the value of AX.

Test MS-DOS-3.10 FAT12 vfs directory ... ok
Test MS-DOS-3.10 FAT16 vfs directory ... ok
Test MS-DOS-3.10 Floppy vfs directory ... ok
Test MS-DOS-3.10 MFS redirection ... ok

Fixes #278