disks: Fix display of MS-DOS 7.0 boot failure message
MS-DOS 7.0 uses IO.SYS failure message on boot failure, but its location
is indicated by an address four bytes before the start of the string.
Previously this seemingly magic sequence had been disabled, but as
MS-DOS 7.0 boot is currently broken, the message is actually shown by
IO.SYS and can be seen to truncated. After comparison with the native
boot block it can be seen that these four bytes are not a magic string,
"\4\3\2\1" as indicated in the fatfs_boot.S commentry but are consistent
with a value being passed for a purpose as yet unknown.