disks: Fix DR-DOS 5 & 6 floppy directory boot
In the old fatfs_boot.S code the stack was set up for its own purposes,
now that loading is done in 32 bit land, the stack wasn't set to a sane
place (0x0000 I seem to remember). This sets the stack to a proper
value and fixes the DR-DOS 5 and 6 floppy directory boot.