simx86: fix horrible bug in PUSHAD and POPAD on x86_64
These instructions were loading the pointer to the TheCPU struct
into the registers, but they used ESI/EDI even on x86_64. As
the result, if the TheCPU struct is located above 4G, the pointers
are invalid, and the memory corruption is what happened.
I made it so RSI/RDI is used on x86_64.
I hope this bug will not manifest in other places...
In principle, simx86 uses only 'lea' insn to load pointers, and
the only "dangerous" is the one with RBX, as it points to TheCPU.
I've only found 2 such places, but if some other techniques are
used for loading pointers, then this bug may re-surface.

Now simx86 jit mode seems really much more stable than it used to be.