avoid stack corruptions when running windows [fixes #25]
This appears to be a win.com bug.
In "standard mode" win.com itself provides the
service to switch to DOS mode (unlike in "enhanced
mode", where this is done by win386.exe's vxd).
When the switch is requested, win.com first switches
to text mode and then loads wswap.exe and dswap.exe.
Unfortunately it reserves the very small stack for
both tasks, about 200 bytes only. Modern video
bioses use more and also enable interrupts during
the modesetting, so the stack data overwrites the
file name in heap that it is about to be run (dswap/wswap).

I've found the following fixes:
1. use winos2. It reserves more stack and is not affected.
2. install windows to the directory with shorter name, for
example "W" instead of "WINDOWS". Since win.com stores
the full file names, with the shorter dir name the fewer heap
will be used, and the vbios will have enough space (yes, the
starvation is just about a couple of bytes!)

This patch does the following:
- disables mouse int10 watcher to save a few stack bytes.
- in case the corruption still happen, it tries to detect it
and repair the filename in memory...

Another option may be to switch stacks when calling real int10
handler, but oh well, we can't work around every bug out there.