windows: switch stacks on vbios calls [#25]
It turns out that the win.com's stack overflow problem is
provoked by display.sys that installs the int10 handler and
calls int10 from it with hw interrupts enabled. This leads
to the heavy stack usage.
With this patch, the alternate stack is used for processing
the vbios calls under windows.
Generic path is unaffected, only windows-specific one is.