msdos_ldt: enlarge initial ldt limit by 4 pages [fixes #287]
It appears krnl286.exe calculates the amount of available LDT
entries by doing 'lsl' and subtracting the allocated descriptor
number from limit. It never allocates above the initial limit,
so to give it more entries, we can enlarge the limit.
Opposite to this, krnl386.exe uses 'lsl' to get the amount of
ldts available for direct writes, and allocates past the limit.