msdos_ldt: remove ldt autoexpanding code
With enlarged limit it seems to be no longer necessary.
Windows works fine, both krnl286.exe and krnl386.exe.
While it may be nice to have an autoexpanding feature, krnl286
can't get the use of it, so we need the large initial limit anyway.
And with such a limit krnl386.exe never writes above it, AFAICS.
If any problem is encountered, we can simply enlarge the limit
by hands.