kbd: return to old good pic hack
bios-assisted hack was a nice try but failed.
If more than one prog hooks kbd interrupt and passes it to bios
(like if GW tsr is loaded), then there are 3 readers in chain, not 2.
And distinguishing the first 2 is not possible, so "identify bios and
return prev value to it" heuristic does not help.