emm: forbid reallocating/freeing OS handle [#225]
Reallocating OS handle resulted in unmapped areas in lowmem_base,
which should never happen. It may be possible to implement the
reallocation of an OS page differently, but so far we don't know
what should it do.
Perhaps win-3.0 does that by mistake.
I was not able to get it to work so far.