sb: remove new_dma_init_count to fix TIM2 game
There is already a dma_init_count that is re-loaded in every dma restart.
Having a duplication of it makes it difficult to re-use previous
counter in commands like 0x1c, because the previous counter can then
be either still in new_dma_init_count or already in dma_init_count.
It is better to not duplicate that field.