int: hook int2f and make a part of it non-revect [fixes #97]
This is needed so that XMS is accessible via "call" rather
than only by "int", because this is how it is called from DPMI.

Note: XMS will be unavailable before the hook as I don't swap
the revect handler. This is hopefully not a problem since the
hooking happens early.
Note: it seems we can't unrevect it the way we do with int21.
Some problems with filesystem then happens. This needs to be
investigated.