dpmi: pass rm regs to post_extender on pm stack
This allows to restore rm regs on switch from rmode, as it was
always done, rather than to delay the restoring up to pmode handling.

Note: this prevents using rm_to_pm_regs() after post_extender()
because the rm regs are already restored at that point. Instead
of adding a similar function that will take the source regs as
an arg, I made post_extender() to translate all regs instead of
returning the mask.