disks: Unimplemented int13 function needs CARRY set
Int13 / 18h - Set Media Type For Format is currently unimplemented, but
it didn't set the CARRY flag. This lead to the caller believing it was
returning valid data in ES:DI when actually it was just the callers
original values. This caused data corruption of the drive number being
formatted under Windows 95. [fixes #276]