sdl: make sure SDL_Quit() is called only once [fixes #345]
SDL API is unfriendly.
We need to make sure SDL_Init() and SDL_Quit() are called
only once and cover all SDL_InitSubSystem()/SDL_QuitSubSystem()
calls. This is problematic because we may have SDL for video,
for sound or for both. So we need to handle every such case
separately...