sound: use MIC instead of LINE for alsa input [#363]
The LINE was used because it is stereo, whereas the MIC input
is mono. Unfortunately windows's sb16 driver enables LINE
pass-through, which makes noise if the microphone is plugged.
By moving microphone to MIC input we avoid the problem.
Fortunately the microphone is hardly stereo, so there is
nothing to lose.