dpmi: allocate pool adjucently unless in vm86 case [#385]
It is not possible to allocate DPMI pool separately because
due to the top-down mmap allocations the pool will be below
mem_base, which exceeds the 1Gb offset that some DOS programs
want. The 1Gb is exceeded because on i386 architecture the
top 1Gb is reserved for the kernel, so the wrap-around 4Gb
boundary always gives an offset over 1Gb.
We can still use the find_hole techique on x86_64 as it will
likely to succeed, but lets not split the code.
In vm86 mode still use find_hole technique as then it has
more chances to succeed than contig allocation at addr 0.