ems: Adjust device driver error exit method
Old versions of DOS (seen with PC-DOS 3.00) did not check the usual
status value in AX at load time. This can lead to the driver being
loaded and maybe because the break address was still set to %cs:0000
the CON device got corrupted. Apparently the correct method to ensure
the driver did not load is to change the driver header attribute to
signify a block device and to set the number of units provided to zero.

Tested with:
	PC-DOS 3.00
	MS-DOS 3.10
	DR-DOS 3.40