msdos: preserve high words of PM regs upon rm calls [fixes #419]
The only exception is long read/write, for which we should
pass full eax by hands.
It used to work by the pure miracle: RMLWORD() macros were
forgetting to mask the value to 0xffff, and DOS code does not
touch the upper words either, so they were preserved via the
real-mode regs. But new revectoring code trashes high words of
some regs, so things broke.
This worked till 5dca733 by another pure luck: msdos.c simply
avoids our revectoring code, so it was unaffected by the
revectoring changes till the post_boot was updated to new logic.
This was quite difficult to fix because the patch to which
git bisect pointed, was innocent, thanks to the 2 miracles above.
Also the long read/write stuff made all my test inconsistent
until I realized that it is a separate bug - eax passing was
forgotten and worked only because it was compensated by the
forgotten masking in RMLWORD().

Quake2/win now works under HX! This time for real.