signal: stop using struct ucontext [fixes #457]
It is no longer exported by glibc:
https://sourceware.org/glibc/wiki/Release/2.26#Removal_of_.27struct_ucontext.27
... unless you include <asm/ucontext.h>, but this will conflict
with older glibc that does export it.
Kernel's struct ucontext subtly differs from glibc's ucontext_t,
but we only need uc_flags and uc_mcontext. uc_flags are at the
beginning of both, and uc_mcontext's location matches but type
differs, so we need to cast it.